+++ src/docs/HISTORY.md
@@ -37,6 +37,9 @@ Rocky Linux's **XDG_SESSION_TYPE** environment variable.
 - Added hotkey to control click (pass) through of window position and clicking.  Works on all platforms.  However, on X11, international keys are not handled and text selection does not work.  The hotkey by default is **Ctrl + t**.
 
 - Fixed pip.sh script to call pip on the local mrv2 install.
+- Added information about environment (desktop, os and kernel) of the machine that was used to build it.
+- Made Build and Running log information tidier.
+- Cleaned up log messages a bit.